business sales data/requirements.txt
discord.py>=2.3,<3.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
resend>=2.0,<3.0
aiohttp>=3.9,<4.0
weasyprint>=62.0,<63.0
